Code refactoring #JC19-360
