+++ apps/vanilla/src/pages/index.ts
@@ -1,7 +1,9 @@
+import { Middleware } from '../namespaces/middleware';
 import { initAnimeTable, initPagination, initQuery } from '../script/init';
-import { checkPermission } from '../script/isAuthorized';
+import { renderHeader } from '../script/renderToUI';
 
+Middleware.checkIsLoggedIn();
 initAnimeTable();
 initPagination();
 initQuery();
-checkPermission();
+renderHeader();
